Renamed metadata to article_index
